backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-marshmallow==1.2.1
flask-migrate==4.0.7
flask-cors==4.0.1
marshmallow-sqlalchemy==1.1.0
marshmallow==3.21.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
cloudinary==1.40.0
openai==1.35.13
bcrypt==4.1.3
email-validator==2.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
